(BSR)[PRO] refactor: move booking tests to TS
