+++ pro/src/screens/Bookings/BookingsRecapTable/components/CellsFormatter/__specs__/BookingDateCell.spec.tsx
@@ -1,21 +1,19 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 
-import BookingDateCell from '../BookingDateCell'
+import BookingDateCell, { BookingDateCellProps } from '../BookingDateCell'
 
-const renderDateCell = props => render(<BookingDateCell {...props} />)
+const renderDateCell = (props: BookingDateCellProps) =>
+  render(<BookingDateCell {...props} />)
 
-describe('bookings date cell', () => {
+describe('BookingDateCell', () => {
   it('should display the date and the time', () => {
-    // Given
     const props = {
       bookingDateTimeIsoString: '2020-04-03T12:00:00+04:00',
     }
 
-    // When
     renderDateCell(props)
 
-    // Then
     expect(screen.getByText('03/04/2020')).toBeInTheDocument()
     expect(screen.getByText('12:00')).toBeInTheDocument()
   })
